rename `plotevokedcomparisondialog` -> `plotevokeddialog`
